use recursion

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -4,7 +4,7 @@
  * @n: number to be printed
  * Return: no return
  */
-void print_number(long int n)
+void print_number(int n)
 {
 	if (n < 0)
 		n *= -1;
